Add `GotenbergAssetRuntime` to avoid passing Builder in context
Passing builder in context is fragile, as it won't work if user
uses any of Twig layout features without passing the full context.

Starting with Twig 4.0, the `include` function won't pass the
full context, to encourage template isolation and composition.

This PR introduce a Twig Runtime, lowering the impact of this extension
on the global scope, and allowing to create a time-limited "scope",
during which the `gotenberg_asset` will work as expected.